ftplugin c: can't rely on SID. use b: variable
i can't define a function as <SID> in ftplugin/ and expect to see it in
my script in plugin/ because the two scripts don't share the same <SID>.
what i can do is use a buffer-local variable.

i've done this before with the C++ ftplugin but apparently forgot to do
the same in the C ftplugin.