Deprecate :IncludeSystem/Local for :Include
This makes the interface easier: only one "Include" command now, instead
of the old two. In turn this means I move the logic of generating a
system/local include to the filtering pipeling. It's a regex work.

Technical details follow: implementing this feature implied
concatenating the logic of selecting system-includes vs. local-includes
into the "substitute" pipeline. If the user returns an empty string from
the input prompt then reimin used to ignores the command even before
processing the substitution pipeline. Now, perform the substituion
pipeline anyway, and then check for the result being empty. This allows
the pipeline to abort the command by retuning an empty string.